Check_input_handler_buffer.cpp add

+++ Config/Srcs/Check_allow_metods.cpp
@@ -6,7 +6,7 @@ bool Config::check_allow_metods(const ConfigHandler::t_params &param, Input_hand
 		return false;
 	else
 	{
-		for (int i = 0; i < param.allow_methods.size(); ++i)
+		for (int i = 0; i < static_cast<int>(param.allow_methods.size()); ++i)
 		{
 			if (param.allow_methods[i] == handlers.getType())
 				return false;
